pyproject.toml
[tool.poetry]
name = "home_interview"
version = "0.1.0"
description = ""
authors = ["Noors <jackmovies01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "2.2"
requests = "^2.27.1"
xmltodict = "^0.12.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"
djangorestframework-camel-case = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
